refactor(map): route Map.razor interop through typed wrappers
37 of the 44 raw module.InvokeAsync sites in Map.razor now go through
the IMap*Js wrappers, dropping the per-site try/catch dance. The
remaining 7 lifecycle-bound sites (initMap, the 6 dispose-time
clear*, dispose) stay bare with explanatory comments -- the
wrappers' MarkDisposed gate would otherwise no-op the teardown.

The InvokeVoidSafeAsync helper is gone; both fire-and-forget
sites in HandleSettingsChanged go through the wrappers now. The
SetGuardZoneAsync wrapper signature took double for lookaheadMin
to match IAppSettings.GuardZoneLookaheadMinutes (was int, now matches).

+++ OnaPlotter/Services/Js/MapControlsJs.cs
@@ -36,7 +36,7 @@ public sealed class MapControlsJs : IMapControlsJs
     public Task SetNightModeAsync(bool enabled)
         => InvokeSafe("setNightMode", enabled);
 
-    public Task SetGuardZoneAsync(double radiusNm, int lookaheadMin, double warningFactor)
+    public Task SetGuardZoneAsync(double radiusNm, double lookaheadMin, double warningFactor)
         => InvokeSafe("setGuardZone", radiusNm, lookaheadMin, warningFactor);
 
     public Task PanToAsync(double lat, double lon)
